Refactored component to separate logic from App.tsx into new storefront component.
