+++ frontend/package.json
@@ -12,7 +12,7 @@
     "@types/react-dom": "^17.0.11",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-scripts": "5.0.0",
+    "react-scripts": "^5.0.0",
     "typescript": "^4.5.5",
     "web-vitals": "^2.1.4"
   },
